pip-requirements.txt
# pip-requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
av==14.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
comfyui-frontend-package==1.16.9
comfyui-workflow-templates==0.1.3
controlnet-aux==0.0.9
diffusers==0.33.1
einops==0.8.1
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
importlib-metadata==8.6.1
jinja2==3.1.6
kornia==0.8.0
kornia-rs==0.1.8
lazy-loader==0.4
markupsafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
pip==25.0
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
py360convert==1.0.3
pycparser==2.22
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentencepiece==0.2.0
setuptools==75.8.0
soundfile==0.13.1
spandrel==0.4.1
sympy==1.13.1
tifffile==2025.3.30
timm<=0.6.7
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchsde==0.2.6
torchvision==0.21.0
tqdm==4.67.1
trampoline==0.1.2
transformers==4.51.3
triton==3.2.0
typing-extensions==4.13.2
urllib3==2.4.0
wheel==0.45.1
yarl==1.20.0
zipp==3.21.0
uvicorn
fastapi
